User TA: add --as-needed to link flags
Add link flag '--as-needed' to avoid adding dependencies onto shared
libraries that are not actually used. This becomes particularly useful
when we build the base user libraries (libutils, libmpa and libmbedtls) as
shared objects. Without the flag, and because all libraries are passed to
the linker by default, all TAs would need to map all the libraries at
runtime resulting in a waste of memory.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>