core: pager: fix tee_pager_unhide_page()
Prior to this patch was tee_pager_unhide_page() searching for a physical
page which was used at a certain page index in an area. What wasn't
checked was that the area in addition to the page index matched. This
leads sometimes unhiding the wrong page which will result in rapid
aborts in succession until the correct page has been handled. With this
patch the area is also checked fixing the problem.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>