core: pager: don't store hidden page pa in table
When a page is hidden the associated pmem->flags is updated with
PMEM_FLAG_HIDDEN. From a pmem it's also possible to derive the physical
address of the page. This makes storing the physical address of a hidden
(and possibly dirty) page redundant. So always store 0 instead of
physical address of hidden pages to simplify.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>