core: pager: add missing tzsram physical pages
init_runtime() adds the physical pages used for paging. Prior to this
patch only pages covered by the memory using by the binary was added. If
TZSRAM is large enough there will remain a range of physical pages after
the "pageable_end" address. With this patch this last range is also added.

Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>