Remove deprecated ldm/stm instructions
Some uses of SP, PC and LR in the register list of Load/Store Multiple
instructions is forbidden (Thumb) or deprecated (ARM) as per the ARM
ARM DDI 0406 C.d. For the LDM instructions, SP should not be in the
list, and the list should not contain both PC and LR. See sections:

 [A8.8.58] LDM/LDMIA/LDMFD (Thumb)
 [A8.8.59] LDM/LDMIA/LDMFD (ARM)
 [A8.8.60] LDMDA/LDMFA
 [A8.8.61] LDMDB/LDMEA
 [A8.8.62] LDMIB/LDMED

For the STM instructions, neither SP nor PC should be in the list. See
sections:

 [A8.8.200] STM (STMIA, STMEA)
 [A8.8.201] STMDA (STMED)
 [A8.8.202] STMDB (STMFD)
 [A8.8.203] STMIB (STMFA)

Clang warns on the deprecated constructs. Use ldr/str instead.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>