core: tee_pager_set_uta_area_attr(): save flags
Prior to this patch is tee_pager_set_uta_area_attr() saving the mattr
bits instead of just the protection bits derived from the flags
parameter. This leads to tee_pager_set_uta_area_attr() updating
permission even when not needed. With this patch is only the effective
protection bits saved in the different struct tee_pager_area which are
updated when changing permissions.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>