core: tee_pager_set_uta_area_attr(): check same context
Prior to this patch it was assumed that only one area could be using a
fobj unless it was shared between multiple context. This isn't true, If
an area happens to span two page tables it will be split into two areas
connected to the same fobj. This patch fixes this by checking that all
areas using a fobj has the context.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>