ta_bin_to_c.py: mark as python3 script
This commit marks the script as a python3 script and update
hex string conversion accordingly.

Moving to python3 requires a change in hex string conversion.
It causes by `zlib.compress()` behaving differently in python2
and python3. It returns a string in python2, and it returns
a bytes object in python3.

With python3 `zlib.compress()` behavior, we can use hex() to
convert bytes[i] of type int into a hex string.

Signed-off-by: Pipat Methavanitpong <pipat.methavanitpong@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>