get_maintainer.py: process patches individually
When given the -m/-merge-check option, get_maintainer.py parses the
Acked-by and Reviewed-by tags that may be found in a patchset or
Github PR. In presence of several patches the tags should apply to
each patch individually, not to the whole patchset as is currently
done. As a result, the script may fail to report some unapproved
changes.

Fix this issue by splitting patchsets into individual patches before
processing.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>