core: boot: relax init sections from DT related resources
Move the initialization sequences that deal with device tree parsing
support out of the text_init and rodata_init sections to relax memory
pressure on them. To do so this change splits init_primary_helper()
in 2 helper functions: generic_init_primary() for resources expected
in the init sections and paged_init_primary() for resources that move
the pageable sections.

Updates inline comments to state generic_boot_init_primary() lies in
the init area, not in the unpaged area as wrongly stated prior this
change.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Etienne Carriere <etienne.carriere@linaro.org> (stm32mp1)