ci: shippable: use our own cache server
Shippable has a maximum build time of 1 hour for Open Source projects.
It is usually sufficient since typical build time is currently around
35 minutes, thanks to caching (ccache). But it can happen that the code
is changed too much by a pull request in which case the cache is
ineffective and build time would exceed 1 hour. When a build times out,
the cache is NOT updated and there is no other choice than to manually
feed the cache with a partial build (by removing some steps in
.shippable.yml) so that the next build can be quicker, wait for the
build to succeed, then restore all the build steps. This is quite
impractical.

This commit replaces the Shippable built-in caching mechanism with our
own cache on a cloud server (currently a Digital Ocean [1] VM).
Automatic upload is triggered after 50 minutes of build time (or at the
end of the build). Therefore, if the build times out we have a
significant amount of data in the cache and restarting the build is
usually all it takes to deal with the issue.

Cache data are transferred using SSH and the cache server is configured
with the public key of the official OP-TEE OS Shippable project (the
one that is used to build pull requests).

In addition to being tolerant to time outs, this new method is usually
faster too, because the archive is streamed directly to/from the server
whereas Shippable creates a local .tar.gz file first.

For all intents and purposes I am documenting the security configuration
of the VM:
- Ubuntu 20.04 with default settings
- Dedicated user account: 'shippable'
- The OP-TEE SSH key [3] is added to ~shippable/.ssh/authorized_keys
- The login shell for user 'shippable' is /usr/sbin/rush [2] with the
  following /etc/rush.rc:

 debug 1

 rule default
   acct on
   limits t10r20
   umask 002
   env - USER LOGNAME HOME PATH
   fall-through

 rule cat-to
   command ^cat >[a-zA-Z0-9_\.-]+$
   chdir /home/shippable
   transform s@(.*)@/usr/bin/bash -c "\\1"@

 rule cat-from
   command ^cat [a-zA-Z0-9_\.-]+$
   chdir /home/shippable
   transform s@(.*)@/usr/bin/bash -c "\\1"@

[1] https://digitalocean.com/
[2] https://www.gnu.org/software/rush/
[3] https://app.shippable.com/subs/github/OP-TEE/settings

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>