Makefile: sort in-tree TAs in lexical order
The Makefile function $(wildcard ...) may return matching files in any
order. Therefore the "make" command is not guaranteed to always produce
the same sequence of commands when building the in-tree TAs. This could
be inconvenient when troubleshooting build errors in different
environments for instance.
In order to produce predictable builds, sort the in-tree TA makefiles in
lexical order.

Suggested-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Jerome Forissier <jerome@forissier.org>