build: ld-option: handle any linker warning as an error
The purpose of the ld-option macro is to detect if the linker supports
a given command line option or not. It does so by invoking the linker
with the option and checking the exit status of the process. Some
options however may not cause an error but only generate a warning
message, and the linker exits with a success status. For example,
'-z unrecognized-option' does cause an error with Clang but triggers a
warning with GCC. As a result, $(call ld-option,-z unrecognized-option)
has a different behavior depending on the compiler.

Address this issue by loooking for the word 'warning' in the linker
output in addition to checking the exit status.

Fixes the following warning when building xtest shared libraries with
GCC:

 path/to/bin/arm-linux-gnueabihf-ld.bfd: warning: -z separate-loadable-segments ignored

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>