ci: travis: download checkpatch.pl from linux-next
The Travis CI script obtains checkpatch.pl from a mirror of the master
branch of the Linux repository. Unfortunately, kernel v5.9 has a buggy
checkpatch.pl [1]:

 Global symbol "$gitroot" requires explicit package name [...]

The fix [2] is available in linux-next already but due to the Linux
release schedule it won't make it to the master branch until the end of
the v5.10-rc1 merge window, which is expected two weeks after the v5.9
release.

Downloading checkpatch.pl from linux-next seems to be a better idea in
general. Of course it might break more often due to the unstable nature
of linux-next, but on the other hand fixes appear much more quickly
than in master. A faster update time is even more beneficial in case
new features are added (such as the --kconfig-prefix option I submitted
some time ago [3]) because new patches have to wait up to several weeks
if they miss the merge window.

Therefore, change the checkpatch URL from linux master to linux-next to
fix the above error immediately and hopefully be more convenient in the
future.

Link: [1] https://travis-ci.org/github/OP-TEE/optee_os/builds/737636665#L264
Link: [2] https://lkml.org/lkml/2020/10/18/167
Link: [3] https://lkml.org/lkml/2020/8/18/212
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>