core: Add ffa_msg_send_req and ffa_msg_send_resp
Add FF-A message handling of FFA_MSG_SEND_DIRECT_REQ_32 and
FFA_MSG_SEND_DIRECT_RESP_32 messages.
The FFA_MSG_SEND_DIRECT_RESP_32 message is used to send a message from
one FF-A endpoint to a second FF-A endpoint. FFA_MSG_SEND_DIRECT_RESP_32
is used to send a response message from the callee to the caller.

A separate abstraction is made (spmc_sp_handler) for handling FF-A
messages received from secure endpoints (SP).
thread_spmc_msg_recv() is still used for handling all FF-A messages
coming from the Normal World. When a FFA_MSG_SEND_DIRECT_REQ_32 message
is received for the Normal World a new kernel thread is started and
spmc_sp_msg_handler() is run. spmc_sp_msg_handler() is used as the main
message handler for handling the FF-A messages coming from SPs.

Whenever a message is sent to the normal world the thread is finished
and the FF-A message is forwarded to the Normal World.

Whenever a SVC exception is received from an SP, the whole context is
saved and executions is resumed in S-EL1. The first 8 register (FF-A
registers x0-x7) contain the FF-A message and are used to handle the
message by the SPMC.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>