shippable: extract Plug-and-Trust library to unversioned directory
The Plug and Trust library archive creates a directory when extracted
that contains the version number. This is fine in general but is
detrimental to our CI script, because the version number needs to
appear 7 times in the script, which is error-prone when upgrading.

Change the extraction command to create /root/se050 instead of
/root/se050-0.0.1, etc. This way the Plug and Trust version number is
only present in the package URL.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jorge.Ramirez-Ortiz <jorge@foundries.io>
Tested-by: Jorge Ramirez-Ortiz <jorge@foundries.io>