core: mm: fix infinite loop in vm_pa2va()
Commit d6ad67f674e5 ("core: mm: change vm_pa2va() to return a virtual
address") moved the call to mobj_get_pa() up in a 'for' loop and added
a 'continue' statement. Moving it was wrong because at this point 'size'
is not yet updated which causes an infinite loop when the PA is not
found.

Move the call back to its original location but keep the 'continue'
which looks correct.

Fixes: d6ad67f674e5 ("core: mm: change vm_pa2va() to return a virtual address")
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>