drivers: csu: allow setting CSU_CSL0
The sentinel detection in the initialization loop for the CSU_CSL<n>
registers is wrong in that is doesn't allow to set the first register,
CSU_CSL0 (when csu_index == 0). Fix the conditional so that it stops
on the sentinel value (-1) but still allows 0 as a valid index.
CSU_CSL0 is used for the PWM peripherals on i.MX6 platforms.

Reported-by: Robert Delien <r.delien@payter.nl>
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Acked-by : Clement Faure <clement.faure@nxp.com>