ci: travis: select Ubuntu 20.04
The default build environment for Travis CI is currently Linux Ubuntu
Xenial (16.04). This is quite an old release (21 April 2016). The latest
release supported by Travis CI is Focal (20.04), released on 23 April
2020. This commit updates to 20.04 by selecting "dist: focal" instead of
keeping the default. The reason is to obtain a more up-to-date codespell
dictionary. For example, it was found that checkpatch detects the
spelling mistake 'becase' on 20.04 but not on 16.04.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>