plat-sam: set sram0 as secure
We are going to use the first 64K of SRAM0 to put the suspend code.
This is mandatory because SDRAM will be put in self refresh mode and
thus won't be accessible. at91bootstrap already sets this zone as
secure but we better enforce it to avoid anybody reading or writing
it.

Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jerome Forissier <jerome@forissier.org>
Signed-off-by: Clément Léger <clement.leger@bootlin.com>