config: Add config options to enable BTI for TEE core and TA's
Branch Target Identification (part of the ARMv8.5 Extensions)
provides a mechanism to limit the set of locations to which
computed branch instructions such as BR or BLR can jump.
To make use of BTI in TEE cores and ldelf on CPU's that
support it, enable the option CFG_CORE_BTI. The option is only
supported for ARM64 cores.

To enable BTI support for TA's and user mode libraries,
enable the option CFG_TA_BTI.

The BTI support is currently not compatible with options
CFG_VIRTUALIZATION and CFG_WITH_PAGER.

To use the option CFG_CORE_BTI, GCC toolchain built with
--enable-standard-branch-protection is needed.

To test with QEMU, use option -cpu max,sve=off.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>
Reviewed-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Jerome Forissier <jerome@forissier.org> (vexpress-qemu_armv8a)