core: kernel: pm: refine suspend/resume sequence order
Change PM generic implementation to call handlers for suspend in the
reserve order in which they were registered and to call handlers for
resume in the opposite order, that is the order in which the handlers
where registered.

This change allows to reuse the driver initialization ordering which is
expected to match device drivers dependencies. This change will be even
more useful when using probe support and dependencies resolution.

The change does not affect the 2 PM registration levels: core and drivers.
Core handlers are still suspends after drivers and resumed before
drivers.

Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Gatien Chevallier <gatien.chevallier@st.com>