ci: azure: fix build error in Buildroot's OpenSSL
After the addition of Buildroot package host-python3-cryptography as a
dependency of optee_client_ext and optee_examples_ext [1], we now have
a build error in the Azure CI "QEMUv8 check" job:

 >>> host-libopenssl 1.1.1l Configuring
 (cd /root/optee_repo_qemu_v8/out-br/build/host-libopenssl-1.1.1l; ...
  ./config ...)
 Operating system: x86_64-whatever-build
 This system (build) is not supported. See file INSTALL for details.

The same command runs OK manually in the Docker image, but not in the
Azure environment. It is due to the fact that Azure sets a number of
environment variables, one of which (${SYSTEM}) conflicts with the
OpenSSL configuration script. Since we don't use this variable in our
script, just unset it.

Link: [1] https://github.com/OP-TEE/build/commit/a3b368f89a1c
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>