core: arm64: suppress text relocations in thread_optee_smc_a64.S
The readjust_pc macro in thread_optee_smc_a64.S relies on loading the
address of a local label (1111f:) from the literal pool, knowing that
this address is modified when relocations are applied. It is a way of
converting PC addresses from the identity map to the "normal" VA space,
i.e. the one to which ASLR has been applied. However, this approach
creates a relocation inside the .text section (TEXTREL) which is
generally undesirable.

Replace the mechanism with a load using the 'adr' instruction, which
obtains an address in the identity map, then add the ASLR offset found
in the boot_mmu_config structure. This gets rid of 8 text relocations.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Sumit Garg <sumit.garg@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>