core: fix race in mobj_reg_shm_inc_map()
Fixes a race in mobj_reg_shm_inc_map() when mapcount is 0. The problem
goes like:

Thread 1 and 2 calls mobj_reg_shm_inc_map() at the same time and
mapcount is 0. Thread 1 takes the lock first and initializes mapcount to
1 and map the mobj etc.

When thread 2 has the lock it discovers that mapcount has been
initialize while it was waiting for the lock.

Prior to this patch we where exiting the function doing nothing more
since the mobj was mapped, but by doing so we'll miss to increase
mapcount.

Fix this by restarting the call to refcount_inc() using a loop.

Fixes: 37a6b717787b ("core: introduce CFG_CORE_DYN_SHM")
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jerome Forissier <jerome@forissier.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>