ldelf: get DT_HASH info from section headers
Obtain the address of the DT_HASH section from the section headers
rather than from the dynamic section. In addition to being simpler,
this method can easily provide the size of the section. It is not
needed in this case but will be useful when introducing support for
DT_GNU_HASH.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>