core: sp: fix sp_dt_get_u64() alignment issue
In the SP DT files 64-bit values are represented by two 32-bit cells.
When loaded into memory, the address of such value might not be 64-bit
aligned. Currently sp_dt_get_u64() simply dereferences a pointer to such
value. This compiles to a 64-bit load instruction, which causes an
alignment fault if the address was not 64-bit aligned.

Replace the direct dereferencing with the fdt64_ld() helper function
which only uses byte loads.

Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>