core: add overflow checks in crypto_aes_ccm_init()
aad_len and payload_len are of the type size_t which has a greater range
than int which is used for the corresponding arguments when passed to
ccm_init(). So to guard against wrapped or truncated values check that
the variables can be in a int first before calling ccm_init().

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>