ci: configure ccache for jobs that use Buildroot
The jobs that use Buildroot are not configured properly to benefit from
ccache when compiling the Builroot packages. Therefore set BR2_CCACHE_DIR
to point to the location that is cached by the CI environment.

The QEMUv8_check_rust job also needs a cache action and should not
change HOME to /root because ccache would default to /root/.cache
instead of /github/home/.cache.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>