optee_os: add alignment to struct internal_aes_gcm_state
In OPTEE_OS, there will be an alignment check on the address
of the `struct internal_aes_gcm_state`, requiring it to be
aligned according to `uint64_t`.
When we compile on the arm32 (qemu) platform, there will be
no problem with this alignment. However, when compiling on
the x86_64 (sim) platform, we will encounter alignment check
issues.
In order to ensure normal operation on the x86_64 (sim) platform
as well, we need to add a forced alignment declaration for
the `struct internal_aes_gcm_state`.
In this way, the `struct internal_aes_gcm_state` can work properly
on both the arm32 (qemu) and x86_64 (sim) platforms.

Signed-off-by: guoshichao <guoshichao@xiaomi.com>