optee_os: make the initcall can work on x64 platform
When building on an x64 system, if we do not add alignment attribute
for struct initcall{},  we observe that the
intervals between the scattered_array sections where
our registered initcalls are located are not directly
adjacent.
Instead, there is an 8-byte padding between each initcall,
which causes issues when optee_os traverses the
scattered_array sections containing the initcalls.
In order to fix this bug, we need to add alignment attribute
for struct initcall{}, and we find out that in x64 platform,
the default alignment is 0x10 for struct initcall{},
so add force alignment(0x10) to struct initcall{}

Signed-off-by: guoshichao <guoshichao@xiaomi.com>